+++ core/php/AbeilleParserD.php
@@ -40,6 +40,7 @@
     include_once __DIR__.'/AbeilleParser-Tuya.php'; // Tuya specific functions
     include_once __DIR__.'/AbeilleParser-Xiaomi.php'; // Xiaomi specific functions
     include_once __DIR__.'/AbeilleParser-Philips.php'; // Philips specific functions
+    include_once __DIR__.'/AbeilleParser-Ikea.php'; // Ikea specific functions
     require_once __DIR__.'/AbeilleModels.php'; // getModelsList()
 
     // Needed for decode8701 and decode8702